extract helius.data
